chore: rename app-config package
