+++ app/src/main/java/com/pereyrarg11/mobile/MainActivity.kt
@@ -14,7 +14,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.pereyrarg11.mobile.core.presentation.navigation.ScreenRoute
 import com.pereyrarg11.mobile.core.presentation.theme.WhiteLabelTheme
-import com.pereyrarg11.mobile.feature.app_config.presentation.AppConfigScreen
+import com.pereyrarg11.mobile.feature.appConfig.presentation.AppConfigScreen
 import com.pereyrarg11.mobile.feature.home.presentation.screen.HomeScreen
 import com.pereyrarg11.mobile.feature.splash.presentation.SplashScreen
 import dagger.hilt.android.AndroidEntryPoint
