+++ app/src/main/java/com/pereyrarg11/mobile/feature/appConfig/presentation/AppConfigScreen.kt
@@ -1,10 +1,10 @@
-package com.pereyrarg11.mobile.feature.app_config.presentation
+package com.pereyrarg11.mobile.feature.appConfig.presentation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.pereyrarg11.mobile.core.presentation.screen.MainScreen
-import com.pereyrarg11.mobile.feature.app_config.presentation.component.AppConfigContent
+import com.pereyrarg11.mobile.feature.appConfig.presentation.component.AppConfigContent
 
 /**
  * It shows the app configuration downloaded from the remote source.
